moved tx to child component
