+++ src/screens/account_details/index.tsx
@@ -21,7 +21,6 @@ const AccountDetails = () => {
   const classes = useStyles();
   const {
     state,
-    loadNextPage,
   } = useAccountDetails();
 
   return (
@@ -72,10 +71,6 @@ const AccountDetails = () => {
             />
             <Transactions
               className={classes.transactions}
-              loadNextPage={loadNextPage}
-              data={state.transactions.data}
-              hasNextPage={state.transactions.hasNextPage}
-              isNextPageLoading={state.transactions.isNextPageLoading}
             />
           </span>
         </LoadAndExist>
